implement available-moves function using a for loop
My first intuition is always to use clojure.core/for on nested structures. In this case,
dealing with the indexes directly is appropriate because it's the information we
need. The other solutions involve double nested reduce-kv which still end up
pulling out the index, but have to do more work to walk the structure and
rebuild the result.