add `optimal-move` function.
Similar to valid-move, optimal-move finds the a next move to play in the game of tic-tac-toe
and returns that added to the end of list of existing moves.

It has 5 steps. win, don't lose, pick center, pick corner, pick any.

Also we opened up the play game function to take a strategy e.g optimal-move.