fix optimal move function
Previously it only worked with a blank board / no moves taken yet. The working
algorithm is more involves picking the spot that touches the most triplets (more
accurately those that the other player doesn't touch either). This is somewhat
involved and due to that i ended up pulling in some repeat functionality that
should be refactored at some point.